Refactor AppName task
It was yelling to much. It is unneded to give a warning status when a customer has at least provided a app name for the main appname setting. The other appname settings in the same config file are optional. I also included looking for the env var for appname. Is common for customers to use it so we should also not be yelling at them when they have not set up a appname in their config file as long as they have an appname set as an env var. I adjusted the unit tests after the changes.